ajustes form_reserva - JS e HTML
