+++ root/pages/acomodacoesold.html
@@ -19,7 +19,7 @@
           <ul>
             <li><a href="/root/index.html">Home</a></li>
             <li><a href="acomodacoes.html" class="active">Acomodações</a></li>
-            <li><a href="formulario.html">RESERVAS</a></li>
+            <li><a href="form_reserva.html">RESERVAS</a></li>
             <li><a href="/root/admin/login_admincss.html">Area do Colaborador</a></li>
             <div class="campologindrop">
               <form action="campologin" method="post">
@@ -39,7 +39,7 @@
       <div id="inicio" class="navigation">
         <div><a href="/root/index.html">Home</a></div>
         <div><a href="acomodacoes.html" class="active">Acomodações</a></div>
-        <div><a href="formulario.html">RESERVAS</a></div>
+        <div><a href="form_reserva.html">RESERVAS</a></div>
         <div><a href="/root/admin/login_admincss.html">Area do Colaborador</a></div>
       </div>
       <div class="campologin">
@@ -113,7 +113,7 @@
               <br /><br />
               <h4>R$ $$$$$</h4>
               <br /><br />
-              <input type="button" value="Reservas!" onclick="location.href='../pages/formulario.html';" />
+              <input type="button" value="Reservas!" onclick="location.href='../pages/form_reserva.html';" />
               <hr />
             </tdRight>
           </tr>
@@ -147,7 +147,7 @@
               <br />
               <h4>R$ $$$</h4>
               <br /><br />
-              <input type="button" value="Reservas!" onclick="location.href='../pages/formulario.html';" />
+              <input type="button" value="Reservas!" onclick="location.href='../pages/form_reserva.html';" />
               <hr />
             </tdRight>
           </tr>
@@ -180,7 +180,7 @@
               <br /><br />
               <h4>R$ $$</h4>
               <br /><br />
-              <input type="button" value="Reservas!" onclick="location.href='../pages/formulario.html';" />
+              <input type="button" value="Reservas!" onclick="location.href='../pages/form_reserva.html';" />
               <br />
               <hr />
             </tdRight>
@@ -213,7 +213,7 @@
       <h3>LINKS ÚTEIS</h3>
       <a href="/root/index.html" class="links">Home</a>
       <a href="acomodacoes.html" class="links">Acomodações</a>
-      <a href="formulario.html" class="links">RESERVAS</a>
+      <a href="form_reserva.html" class="links">RESERVAS</a>
     </div>
     <!--/RodapéFooter-->
   </footer>
