+++ root/samples/template.html
@@ -16,7 +16,7 @@
             <ul class="navigation">
                 <li><a href="index.html" class="active">Home</a></li>
                 <li><a href="pages/acomodacoes.html">Acomodações</a></li>
-                <li><a href="pages/formulario.html">RESERVAS</a></li>
+                <li><a href="pages/form_reserva.html">RESERVAS</a></li>
             </ul>
             <div class="campologin">
                 <a href="pages/sobrenos.html" class="links">Area do Colaborador</a>
